Service para producto financiero

+++ src/main/java/com/prueba/tecnica/financiero/repository/ProductoFinancieroRepository.java
@@ -4,8 +4,8 @@ import com.prueba.tecnica.financiero.model.ProductoFinanciero;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Repository
-public interface ProductoFinancieroRepository extends JpaRepository<ProductoFinanciero, BigDecimal> {
+public interface ProductoFinancieroRepository extends JpaRepository<ProductoFinanciero, BigInteger> {
 }
